requirements.txt
flask==2.1.1
python-dotenv==0.19
flasgger==0.9.5
SQLAlchemy==1.4.35
psycopg2-binary==2.9.3
redis==4.2.2
passlib==1.7.4
flask-jwt-extended==4.3.1
pydantic==1.9.0
pytest==7.1.2
